# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "https://archimed.justcode.am/",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/get_ApplicationUse": {
            "get": {
                "tags": [
                    "About_application"
                ],
                "summary": "get_ApplicationUse",
                "description": "Правила пользования приложением",
                "operationId": "get_ApplicationUse",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Правила пользования приложением",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/get_PrivacyPolicy": {
            "get": {
                "tags": [
                    "About_application"
                ],
                "summary": "get_PrivacyPolicy",
                "description": "Политика конфиденциальности",
                "operationId": "get_PrivacyPolicy",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Политика конфиденциальности",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/get_TermsOfService": {
            "get": {
                "tags": [
                    "About_application"
                ],
                "summary": "get_TermsOfService",
                "description": "Условия предоставления услуг",
                "operationId": "get_TermsOfService",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Условия предоставления услуг",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/get_analise_and_service": {
            "get": {
                "tags": [
                    "Analise and service"
                ],
                "summary": "get_basket",
                "description": "Получение данных из цветка анализов",
                "operationId": "get_analise_and_service",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Получение данных из цветка анализов",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
...
        "/api/get_basket": {
            "get": {
                "tags": [
                    "Basket"
                ],
                "summary": "get_basket",
                "description": "Получаем список моей корзини если нужен один экзепляр отправлаем  /api/get_basket/2 ОТПРАВЛАЕМ TOKEN",
                "operationId": "get_basket",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Basket data",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}